have manual trigger use acknowledge
